pyproject.toml
[project]
name = "tds-proj-2"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "beautifulsoup4>=4.13.4",
    "duckdb>=1.3.2",
    "fastapi>=0.116.1",
    "fastparquet>=2024.11.0",
    "google-generativeai>=0.8.5",
    "html5lib>=1.1",
    "lxml>=6.0.0",
    "matplotlib>=3.10.5",
    "numpy>=2.3.2",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "pillow>=11.3.0",
    "pyarrow>=21.0.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.4",
    "scikit-learn>=1.7.1",
    "scipy>=1.16.1",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.5",
    "uvicorn>=0.35.0",
    "xlrd>=2.0.2",
]
